requirements.txt
# requirements.txt

pygame-ce>=2.4.0
pygame-gui>=0.6.9
numpy>=1.21.0
numba>=0.55.0
scipy>=1.7.0
Pillow>=9.0.0